finalizado

+++ src/review/review.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Body, Param, Delete, Req, UseGuards } from '@nestjs/common';
+import { Controller, Post, Body, Param, Delete, Req, UseGuards, ForbiddenException, NotFoundException } from '@nestjs/common';
 import { ReviewService } from './review.service';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { OAuthGuard } from '../guards/oauth.guard';
@@ -14,8 +14,15 @@ export class ReviewController {
     return this.reviewService.create(createReviewDto, req.user);
   }
 
+  @UseGuards(OAuthGuard)
   @Delete(':id')
-  delete(@Param('id') id: number) {
+  async delete(@Param('id') id: number, @Req() req: any) {
+    const review = await this.reviewService.findOne(+id)
+
+    if(!review) throw new NotFoundException("Review not found")
+
+    if(review.userEmail !== req.user.email) throw new ForbiddenException("You can't delete this review")
+
     return this.reviewService.delete(+id);
   }
 }
